% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ip_data.R
\name{save_ip_data}
\alias{save_ip_data}
\title{Save IP Data to CSV}
\usage{
save_ip_data(ip_addresses, api_key, base_url, output_file = "IP2location.csv")
}
\arguments{
\item{ip_addresses}{A vector of IP addresses.}

\item{api_key}{Your IP2Location.io API key.}

\item{base_url}{The base URL for the IP2Location.io API.}

\item{output_file}{The file path where to save the CSV (default is "IP2location.csv").}
}
\value{
None (data is written to a CSV).
}
\description{
This function retrieves information for multiple IP addresses using the IP2Location.io API,
processes the data, and saves it to a CSV file. The function also handles missing values by assigning NA
for any field not returned by a given API plan version.
}
